feat(email): replace sendgrid.ts with typed sendgrid.tsx (PR β task 6)
